+++ test/zoi/struct_test.exs
@@ -1,0 +1,30 @@
+defmodule Zoi.StructTest do
+  use ExUnit.Case, async: true
+  doctest Zoi.Struct
+
+  describe "enforce_keys/1" do
+    test "return correct enforce_keys" do
+      schema =
+        Zoi.struct(__MODULE__, %{
+          name: Zoi.string(),
+          age: Zoi.integer() |> Zoi.default(0) |> Zoi.optional(),
+          email: Zoi.string()
+        })
+
+      assert Zoi.Struct.enforce_keys(schema) == [:email, :name]
+    end
+  end
+
+  describe "struct_fields/1" do
+    test "return correct struct_fields" do
+      schema =
+        Zoi.struct(__MODULE__, %{
+          name: Zoi.string() |> Zoi.required(),
+          age: Zoi.integer() |> Zoi.default(0),
+          email: Zoi.string()
+        })
+
+      assert Zoi.Struct.struct_fields(schema) == [:name, {:age, 0}, :email]
+    end
+  end
+end